Prometheus sidecar memory optimization  (#769)
Don't start telegraf, mdsd, and fluent-bit in the prometheus sidecar if it has no work to do (monitor_kubernetes_pods = false and no OSM namespaces to scrape). This part is just a resource-usage optimization.

Adding the newly created environment variables in a file as adding them to bashrc makes it inaccessible if being run in a non-interactive environment. This happens in case of livenessprobe.sh.